# docs/api/cotsel-dashboard-gateway.openapi.yml
openapi: 3.1.0
info:
  title: Cotsel Dashboard Gateway API
  version: 0.1.0
  summary: Contract-first admin/operator gateway for Cotsel-Dash.
  description: |
    Contract-first API specification for the Cotsel-Dash admin/operator dashboard to interact with Cotsel.

    Discovery outcome:
    - Dedicated in-repo dashboard gateway runtime: `gateway/` Express + TypeScript service.
    - Existing repo auth/session surface: `auth` service issues Bearer session tokens.
    - Existing repo governance source of truth: `contracts/src/AgroasysEscrow.sol`, `sdk/src/modules/adminSDK.ts`, and governance runbooks.
    - Existing repo compliance source of truth: policy/runbooks only; no dedicated compliance HTTP service exists yet.

    This specification is therefore contract-first for the dashboard gateway, but it is grounded in existing repo truth.

    Gateway authorization policy:
    - Auth uses existing auth-service Bearer sessions.
    - Gateway role mapping is derived from current auth role `admin`, which maps to `operator:read` and `operator:write`.
    - All mutation endpoints additionally require `GATEWAY_ENABLE_MUTATIONS=true` and membership in `GATEWAY_WRITE_ALLOWLIST`.
...
servers:
  - url: http://127.0.0.1:3600/api/dashboard-gateway/v1
    description: Local contract-first gateway base URL.
...
security:
  - sessionBearer: []
paths:
...
  /settlement/handoffs:
    post:
      tags: [Settlement]
      summary: Ingest a platform settlement handoff
      description: |
        Service-authenticated ingress contract for platform backends to register a settlement handoff with Cotsel.
        This route is not intended for operator sessions; it is authenticated using service-auth HMAC headers and
        requires an `Idempotency-Key` header.
      operationId: createSettlementHandoff
      security: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementHandoffCreateRequest'
      responses:
        '202':
          description: Settlement handoff accepted into the Cotsel orchestration boundary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementHandoffResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /settlement/handoffs/{handoffId}/execution-events:
    get:
      tags: [Settlement]
      summary: List settlement execution events for a handoff
      operationId: listSettlementExecutionEvents
      security: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/SettlementHandoffId'
      responses:
        '200':
          description: Settlement execution events for the handoff.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementExecutionEventListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Settlement]
      summary: Ingest execution or reconciliation truth for a settlement handoff
      description: |
        Service-authenticated mutation for authoritative execution and reconciliation truth coming back into the
        Cotsel gateway boundary. Every accepted event produces a durable callback-delivery record for the configured
        platform backend when callback delivery is enabled.
      operationId: createSettlementExecutionEvent
      security: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/SettlementHandoffId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementExecutionEventCreateRequest'
      responses:
        '202':
          description: Execution or reconciliation event accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementExecutionEventMutationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /governance/status:
    get:
      tags: [Governance]
      summary: Current protocol governance status
      operationId: getGovernanceStatus
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: Snapshot of on-chain governance-relevant status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /governance/actions:
    get:
      tags: [Governance]
      summary: List recent governance actions
      operationId: listGovernanceActions
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/GovernanceActionCategory'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/GovernanceActionStatus'
        - name: tradeId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Governance action history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceActionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /governance/actions/{actionId}:
    get:
      tags: [Governance]
      summary: Get governance action by id
      operationId: getGovernanceAction
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Governance action detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceActionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: session-id
      description: |
        Session bearer token issued by the in-repo auth service.

        Production path:
        - trusted upstream session exchange via `/api/auth/v1/session/exchange/agroasys`

        Compatibility path:
        - `/api/auth/v1/login` after explicit wallet-signature login when that legacy flow is enabled

        Repo evidence:
        - `auth/src/api/controller.ts`
        - `auth/src/middleware/middleware.ts`
        - `auth/README.md`

        The gateway must resolve the session subject and enforce endpoint-level role policy. Governance and compliance mutations require operator/admin authorization.
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Client-generated request identifier for end-to-end traceability. Gateway should generate one if absent.
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Stable correlation identifier linking dashboard actions to downstream service logs, trade events, and incident artifacts.
...
    SettlementHandoffId:
      name: handoffId
      in: path
      required: true
      schema:
        type: string
      description: Cotsel trade identifier used across chain events, indexer evidence, and operator workflows.
  responses:
...
    UnauthorizedError:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated subject lacks required role or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Request conflicts with current state or prior idempotent request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        timestamp:
          type: string
          format: date-time
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [AUTH_UNAUTHORIZED, GOV_STATE_CONFLICT, CMP_PROVIDER_UNAVAILABLE]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
        requestId:
          type: string
...
    SettlementExecutionStatus:
      type: string
      enum: [pending, accepted, queued, submitted, confirmed, failed, rejected]
    SettlementReconciliationStatus:
      type: string
      enum: [pending, matched, drift, unavailable]
    SettlementCallbackStatus:
      type: string
      enum: [pending, delivered, failed, dead_letter, disabled]
    SettlementEventType:
      type: string
      enum: [accepted, queued, submitted, confirmed, failed, rejected, reconciled, drift_detected]
    SettlementHandoff:
      type: object
      required:
        - handoffId
        - platformId
        - platformHandoffId
        - tradeId
        - phase
        - settlementChannel
        - displayCurrency
        - displayAmount
        - assetSymbol
        - assetAmount
        - ricardianHash
        - externalReference
        - metadata
        - executionStatus
        - reconciliationStatus
        - callbackStatus
        - providerStatus
        - txHash
        - latestEventType
        - latestEventDetail
        - latestEventAt
        - callbackDeliveredAt
        - requestId
        - sourceApiKeyId
        - createdAt
        - updatedAt
      properties:
        handoffId:
          type: string
        platformId:
          type: string
        platformHandoffId:
          type: string
        tradeId:
          type: string
        phase:
          type: string
        settlementChannel:
          type: string
        displayCurrency:
          type: string
        displayAmount:
          type: number
        assetSymbol:
          type: [string, 'null']
        assetAmount:
          type: [number, 'null']
        ricardianHash:
          type: [string, 'null']
        externalReference:
          type: [string, 'null']
        metadata:
          type: object
          additionalProperties: true
        executionStatus:
          $ref: '#/components/schemas/SettlementExecutionStatus'
        reconciliationStatus:
          $ref: '#/components/schemas/SettlementReconciliationStatus'
        callbackStatus:
          $ref: '#/components/schemas/SettlementCallbackStatus'
        providerStatus:
          type: [string, 'null']
        txHash:
          type: [string, 'null']
          description: Canonical EVM transaction hash when available.
        latestEventType:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/SettlementEventType'
        latestEventDetail:
          type: [string, 'null']
        latestEventAt:
          type: [string, 'null']
          format: date-time
        callbackDeliveredAt:
          type: [string, 'null']
          format: date-time
        requestId:
          type: string
        sourceApiKeyId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SettlementExecutionEvent:
      type: object
      required:
        - eventId
        - handoffId
        - eventType
        - executionStatus
        - reconciliationStatus
        - providerStatus
        - txHash
        - detail
        - metadata
        - observedAt
        - requestId
        - sourceApiKeyId
        - createdAt
      properties:
        eventId:
          type: string
        handoffId:
          type: string
        eventType:
          $ref: '#/components/schemas/SettlementEventType'
        executionStatus:
          $ref: '#/components/schemas/SettlementExecutionStatus'
        reconciliationStatus:
          $ref: '#/components/schemas/SettlementReconciliationStatus'
        providerStatus:
          type: [string, 'null']
        txHash:
          type: [string, 'null']
          description: Canonical EVM transaction hash when available.
        detail:
          type: [string, 'null']
        metadata:
          type: object
          additionalProperties: true
        observedAt:
          type: string
          format: date-time
        requestId:
          type: string
        sourceApiKeyId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    SettlementHandoffCreateRequest:
      type: object
      required:
        - platformId
        - platformHandoffId
        - tradeId
        - phase
        - settlementChannel
        - displayCurrency
        - displayAmount
      properties:
        platformId:
          type: string
        platformHandoffId:
          type: string
        tradeId:
          type: string
        phase:
          type: string
        settlementChannel:
          type: string
        displayCurrency:
          type: string
        displayAmount:
          type: number
        assetSymbol:
          type: string
        assetAmount:
          type: number
        ricardianHash:
          type: string
        externalReference:
          type: string
        metadata:
          type: object
          additionalProperties: true
    SettlementExecutionEventCreateRequest:
      type: object
      required:
        - eventType
        - executionStatus
        - reconciliationStatus
        - observedAt
      properties:
        eventType:
          $ref: '#/components/schemas/SettlementEventType'
        executionStatus:
          $ref: '#/components/schemas/SettlementExecutionStatus'
        reconciliationStatus:
          $ref: '#/components/schemas/SettlementReconciliationStatus'
        providerStatus:
          type: string
        txHash:
          type: string
          description: Canonical EVM transaction hash when available.
        detail:
          type: string
        metadata:
          type: object
          additionalProperties: true
        observedAt:
          type: string
          format: date-time
    SettlementHandoffResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/SettlementHandoff'
        timestamp:
          type: string
          format: date-time
    SettlementExecutionEventMutationPayload:
      type: object
      required: [handoff, event, callbackDelivery]
      properties:
        handoff:
          $ref: '#/components/schemas/SettlementHandoff'
        event:
          $ref: '#/components/schemas/SettlementExecutionEvent'
        callbackDelivery:
          $ref: '#/components/schemas/SettlementCallbackDelivery'
    SettlementExecutionEventMutationResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/SettlementExecutionEventMutationPayload'
        timestamp:
          type: string
          format: date-time
    SettlementExecutionEventListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/SettlementExecutionEvent'
        timestamp:
          type: string
          format: date-time
...
    GovernanceActionCategory:
      type: string
      enum:
        - pause
        - unpause
        - claims_pause
        - claims_unpause
        - treasury_sweep
        - treasury_payout_receiver_update
        - oracle_disable_emergency
        - oracle_update
    GovernanceActionStatus:
      type: string
      enum:
        - requested
        - submitted
        - prepared
        - broadcast_pending_verification
        - broadcast
        - pending_approvals
        - approved
        - executed
        - cancelled
        - expired
        - stale
        - failed
...
    GovernanceStatus:
      type: object
      required:
        - paused
        - claimsPaused
        - oracleActive
        - oracleAddress
        - treasuryAddress
        - treasuryPayoutAddress
        - governanceApprovalsRequired
        - governanceTimelockSeconds
      properties:
        paused:
          type: boolean
        claimsPaused:
          type: boolean
        oracleActive:
          type: boolean
        oracleAddress:
          type: string
        treasuryAddress:
          type: string
        treasuryPayoutAddress:
          type: string
        governanceApprovalsRequired:
          type: integer
          minimum: 1
        governanceTimelockSeconds:
          type: integer
          minimum: 0
          description: Derived from on-chain `governanceTimelock`.
        requiredAdminCount:
          type: integer
          minimum: 1
          description: Current required admin quorum for unpause flow.
        hasActiveUnpauseProposal:
          type: boolean
        activeUnpauseApprovals:
          type: integer
          minimum: 0
        activeOracleProposalIds:
          type: array
          items:
            type: integer
            minimum: 0
        activeTreasuryPayoutReceiverProposalIds:
          type: array
          items:
            type: integer
            minimum: 0
    GovernanceStatusResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/GovernanceStatus'
        timestamp:
          type: string
          format: date-time
...
    GovernanceActionResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/GovernanceAction'
        timestamp:
          type: string
          format: date-time
    GovernanceActionListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/GovernanceAction'
            nextCursor:
              type: string
              nullable: true
        timestamp:
          type: string
          format: date-time
...